TINYGL: Move more TexelBuffer code into texelbuffer.cpp
