+++ graphics/tinygl/texelbuffer.h
@@ -23,7 +23,6 @@
 #define GRAPHICS_TEXELBUFFER_H
 
 #include "graphics/pixelformat.h"
-#include "graphics/tinygl/colormasks.h"
 
 namespace TinyGL {
 
@@ -48,71 +47,8 @@ protected:
 	float _widthRatio, _heightRatio;
 };
 
-class BaseNearestTexelBuffer : public TexelBuffer {
-public:
-	BaseNearestTexelBuffer(const byte *buf, const Graphics::PixelFormat &format, uint width, uint height, uint textureSize);
-	~BaseNearestTexelBuffer();
-
-protected:
-	byte *_buf;
-	Graphics::PixelFormat _format;
-};
-
-template<uint Format, uint Type>
-class NearestTexelBuffer final : public BaseNearestTexelBuffer {
-public:
-	NearestTexelBuffer(const byte *buf, const Graphics::PixelFormat &format, uint width, uint height, uint textureSize)
-	  : BaseNearestTexelBuffer(buf, format, width, height, textureSize) {}
-
-protected:
-	void getARGBAt(
-		uint pixel,
-		uint, uint,
-		uint8 &a, uint8 &r, uint8 &g, uint8 &b
-	) const override {
-		Pixel col = *(((const Pixel *)_buf) + pixel);
-		_format.colorToARGBT<ColorMask>(col, a, r, g, b);
-	}
-
-	typedef ColorMasks<Format, Type> ColorMask;
-	typedef typename ColorMask::PixelType Pixel;
-};
-
-template<>
-class NearestTexelBuffer<TGL_RGB, TGL_UNSIGNED_BYTE> final : public BaseNearestTexelBuffer {
-public:
-	NearestTexelBuffer(const byte *buf, const Graphics::PixelFormat &format, uint width, uint height, uint textureSize)
-	  : BaseNearestTexelBuffer(buf, format, width, height, textureSize) {}
-
-protected:
-	void getARGBAt(
-		uint pixel,
-		uint, uint,
-		uint8 &a, uint8 &r, uint8 &g, uint8 &b
-	) const override {
-		byte *col = _buf + (pixel * 3);
-		a = 0xff;
-		r = col[0];
-		g = col[1];
-		b = col[2];
-	}
-};
-
-class BilinearTexelBuffer : public TexelBuffer {
-public:
-	BilinearTexelBuffer(byte *buf, const Graphics::PixelFormat &format, uint width, uint height, uint textureSize);
-	~BilinearTexelBuffer();
-
-protected:
-	void getARGBAt(
-		uint pixel,
-		uint ds, uint dt,
-		uint8 &a, uint8 &r, uint8 &g, uint8 &b
-	) const override;
-
-private:
-	uint32 *_texels;
-};
+TexelBuffer *createNearestTexelBuffer(const byte *buf, const Graphics::PixelFormat &pf, uint format, uint type, uint width, uint height, uint textureSize);
+TexelBuffer *createBilinearTexelBuffer(byte *buf, const Graphics::PixelFormat &pf, uint format, uint type, uint width, uint height, uint textureSize);
 
 } // end of namespace TinyGL
